requirements.txt
anthropic>=0.25.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
python-dotenv>=1.0.0
requests>=2.31.0
lxml>=5.0.0
